test_darray: use size_t for the only loop iterator

+++ libnopegl/src/test_darray.c
@@ -57,7 +57,7 @@ int main(void)
     count = ngli_darray_count(&darray);
     ngli_assert(count == 0);
 
-    for (int i = 0; i < 1000; i++) {
+    for (size_t i = 0; i < 1000; i++) {
         element = ngli_darray_push(&darray, NULL);
         ngli_assert(element);
     }
